refactor, got rid of Atomic

+++ src/main/java/kz/job4j/dreamjob/controller/CandidateController.java
@@ -7,8 +7,6 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.concurrent.atomic.AtomicReference;
-
 @Controller
 @ThreadSafe
 @RequestMapping("/candidates")
@@ -19,6 +17,7 @@ public class CandidateController {
     private static final String NOT_FOUND_PAGE = "errors/404";
     private static final String NOT_FOUND_MESSAGE = "Резюме с указанным идентификатором не найдено";
     private static final String MESSAGE_ATTRIBUTE = "message";
+    private static final String CANDIDATE_ATTRIBUTE = "candidate";
 
     public CandidateController(CandidateService candidateService) {
         this.candidateService = candidateService;
@@ -43,37 +42,33 @@ public class CandidateController {
 
     @GetMapping("/{id}")
     public String getById(Model model, @PathVariable int id) {
-        AtomicReference<String> page = new AtomicReference<>(NOT_FOUND_PAGE);
-        candidateService.findById(id).ifPresentOrElse(
-                candidate -> {
-                    model.addAttribute("candidate", candidate);
-                    page.set("candidates/one");
-                }, () -> model.addAttribute(MESSAGE_ATTRIBUTE, NOT_FOUND_MESSAGE)
-        );
-        return page.get();
+        var candidateOptional = candidateService.findById(id);
+        if (candidateOptional.isEmpty()) {
+            model.addAttribute(MESSAGE_ATTRIBUTE, NOT_FOUND_MESSAGE);
+            return NOT_FOUND_PAGE;
+        }
+        model.addAttribute(CANDIDATE_ATTRIBUTE, candidateOptional.get());
+        return "candidates/one";
     }
 
     @PostMapping("/update")
     public String update(@ModelAttribute Candidate candidate, Model model) {
-        AtomicReference<String> page = new AtomicReference<>(NOT_FOUND_PAGE);
-        candidateService.findById(candidate.getId()).ifPresentOrElse(
-                candidateToUpdate -> {
-                    candidateService.update(candidate);
-                    page.set(REDIRECT_CANDIDATES);
-                }, () -> model.addAttribute(MESSAGE_ATTRIBUTE, NOT_FOUND_MESSAGE)
-        );
-        return page.get();
+        var isUpdated = candidateService.update(candidate);
+        if (!isUpdated) {
+            model.addAttribute(MESSAGE_ATTRIBUTE, NOT_FOUND_MESSAGE);
+            return NOT_FOUND_PAGE;
+        }
+        return REDIRECT_CANDIDATES;
     }
 
     @GetMapping("/delete/{id}")
     public String delete(Model model, @PathVariable int id) {
-        AtomicReference<String> page = new AtomicReference<>(NOT_FOUND_PAGE);
-        candidateService.findById(id).ifPresentOrElse(
-                candidateToDelete -> {
-                    candidateService.deleteById(id);
-                    page.set(REDIRECT_CANDIDATES);
-                }, () -> model.addAttribute(MESSAGE_ATTRIBUTE, NOT_FOUND_MESSAGE)
-        );
-        return page.get();
+        var candidateToDelete = candidateService.findById(id);
+        if (candidateToDelete.isEmpty()) {
+            model.addAttribute(MESSAGE_ATTRIBUTE, NOT_FOUND_MESSAGE);
+            return NOT_FOUND_PAGE;
+        }
+        candidateService.deleteById(id);
+        return REDIRECT_CANDIDATES;
     }
 }
